uservars: change type data field from char* to uint64_t
Store user variable types as uint64_t instead of a string. This enables
the possibility of definint flags. Flags are used to delete a user
variable (to destinguish between setting it empty or removing it) and to
mark variables with special properties. This will be used in a following
commit to define global user variables.

Bits 0..31 are reserved for standard datatype definitions,
bits 32..48 are reserved for user defined datatypes or flags,
bits 49..63 are reserved for internal flags.

Signed-off-by: Andreas Reichel <andreas.reichel.ext@siemens.com>
[Jan: coding style fixes]
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>