uservars: Implement USERVAR_TYPE_GLOBAL flag
If the above flag is set in the type field of a user variable,
bgenv_set_uservar_global is called, which iterates all environments
and stores the given variable globally.

If the flag is absent, only the environment specified by the context
handle gets affected.

The flag is set per default if not otherwise specified.

Signed-off-by: Andreas Reichel <andreas.reichel.ext@siemens.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>